Added Cropper tests to test suite
